feat: added mongo and kc-adapter
Mongo:
Database and user are created when the service is initialized. When connecting to mongo from backend services, make use of the following envs: MONGO_DB, MONGO_DB_USER, MONGO_DB_PASSWORD. A database and a user are gonna be common for all backend services.

Keycloak adapter:
The same adapter is supposed to be reused by different backend services, therefore it was moved to a separate location. However, you may not COPY locations from parent directories relative to where a Dockerfile is located. I tried to experiment with build context, but to no avail. For now the solution I came up with is to simply publish it to the public npm registry and then install it upon building like any other dependancy specified in package.json.
See https://www.npmjs.com/package/kc-adapter and echo service for example usage.